fixed render from file upload
